Extract session from base.py
