+++ scripts/exit_node/run_exit_node.py
@@ -16,7 +16,7 @@ from ipv8.taskmanager import TaskManager
 
 from tribler.core import notifications
 from tribler.core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
-from tribler.core.components.base import Session
+from tribler.core.components.session import Session
 from tribler.core.components.ipv8.ipv8_component import Ipv8Component
 from tribler.core.components.key.key_component import KeyComponent
 from tribler.core.components.resource_monitor.resource_monitor_component import ResourceMonitorComponent
