+++ src/tribler/core/components/key/tests/test_key_component.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tribler.core.components.base import Session
+from tribler.core.components.session import Session
 from tribler.core.components.key.key_component import KeyComponent
