+++ src/tribler/core/components/libtorrent/tests/test_libtorrent_component.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tribler.core.components.base import Session
+from tribler.core.components.session import Session
 from tribler.core.components.key.key_component import KeyComponent
 from tribler.core.components.libtorrent.libtorrent_component import LibtorrentComponent
 from tribler.core.components.socks_servers.socks_servers_component import SocksServersComponent
