+++ src/tribler/core/components/metadata_store/tests/test_metadata_store_component.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tribler.core.components.base import Session
+from tribler.core.components.session import Session
 from tribler.core.components.ipv8.ipv8_component import Ipv8Component
 from tribler.core.components.key.key_component import KeyComponent
 from tribler.core.components.metadata_store.metadata_store_component import MetadataStoreComponent
