+++ src/tribler/core/components/payout/tests/test_payout_component.py
@@ -2,7 +2,7 @@
 import pytest
 
 from tribler.core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
-from tribler.core.components.base import Session
+from tribler.core.components.session import Session
 from tribler.core.components.ipv8.ipv8_component import Ipv8Component
 from tribler.core.components.key.key_component import KeyComponent
 from tribler.core.components.payout.payout_component import PayoutComponent
