+++ src/tribler/core/components/reporter/tests/test_reporter_component.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tribler.core.components.base import Session
+from tribler.core.components.session import Session
 from tribler.core.components.key.key_component import KeyComponent
 from tribler.core.components.reporter.reporter_component import ReporterComponent
