+++ src/tribler/core/components/resource_monitor/tests/test_resource_monitor_component.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tribler.core.components.base import Session
+from tribler.core.components.session import Session
 from tribler.core.components.key.key_component import KeyComponent
 from tribler.core.components.resource_monitor.resource_monitor_component import ResourceMonitorComponent
