+++ src/tribler/core/components/tests/test_base_component.py
@@ -1,7 +1,7 @@
 import pytest
 
-from tribler.core.components.base import Component, MissedDependency, NoneComponent, Session, SessionError
-
+from tribler.core.components.base import Component, MissedDependency, NoneComponent
+from tribler.core.components.session import Session, SessionError
 
 
 class ComponentTestException(Exception):
