+++ src/tribler/core/utilities/tiny_tribler_service.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import List, Optional
 
 from tribler.core.components.base import Component
-from tribler.core.start_core import Session
+from tribler.core.components.session import Session
 from tribler.core.utilities.osutils import get_root_state_directory
 from tribler.core.utilities.process_checker import ProcessChecker
